improve parseJSONArray function and add constants file
