+++ client/src/features/admin/AddSessionEnvironmentButton.tsx
@@ -35,7 +35,8 @@ import SessionEnvironmentFormContent, {
   SessionEnvironmentForm,
 } from "./SessionEnvironmentFormContent";
 import { useAddSessionEnvironmentMutation } from "./adminSessions.api";
-import { safeParse } from "../sessionsV2/session.utils";
+import { safeParseJSONArray } from "../sessionsV2/session.utils";
+import { ErrorAlert } from "../../components/Alert";
 
 export default function AddSessionEnvironmentButton() {
   const [isOpen, setIsOpen] = useState(false);
@@ -64,6 +65,7 @@ function AddSessionEnvironmentModal({
   toggle,
 }: AddSessionEnvironmentModalProps) {
   const [addSessionEnvironment, result] = useAddSessionEnvironmentMutation();
+  const [submitError, setSubmitError] = useState(false);
 
   const {
     control,
@@ -80,23 +82,27 @@ function AddSessionEnvironmentModal({
   });
   const onSubmit = useCallback(
     (data: SessionEnvironmentForm) => {
-      addSessionEnvironment({
-        container_image: data.container_image,
-        name: data.name,
-        default_url: data.default_url.trim() ? data.default_url : undefined,
-        description: data.description.trim() ? data.description : undefined,
-        port: data.port ?? undefined,
-        working_directory: data.working_directory.trim()
-          ? data.working_directory
-          : undefined,
-        mount_directory: data.mount_directory.trim()
-          ? data.working_directory
-          : undefined,
-        uid: data.uid ?? undefined,
-        gid: data.gid ?? undefined,
-        command: safeParse(data.command),
-        args: safeParse(data.args),
-      });
+      const commandParsed = safeParseJSONArray(data.command);
+      const argsParsed = safeParseJSONArray(data.args);
+      if (commandParsed === false || argsParsed === false) setSubmitError(true);
+      else
+        addSessionEnvironment({
+          container_image: data.container_image,
+          name: data.name,
+          default_url: data.default_url.trim() ? data.default_url : undefined,
+          description: data.description.trim() ? data.description : undefined,
+          port: data.port ?? undefined,
+          working_directory: data.working_directory.trim()
+            ? data.working_directory
+            : undefined,
+          mount_directory: data.mount_directory.trim()
+            ? data.working_directory
+            : undefined,
+          uid: data.uid ?? undefined,
+          gid: data.gid ?? undefined,
+          command: commandParsed,
+          args: argsParsed,
+        });
     },
     [addSessionEnvironment]
   );
@@ -132,7 +138,11 @@ function AddSessionEnvironmentModal({
         <ModalHeader toggle={toggle}>Add session environment</ModalHeader>
         <ModalBody>
           {result.error && <RtkErrorAlert error={result.error} />}
-
+          {submitError && (
+            <ErrorAlert>
+              The command or arguments are not a valid JSON array
+            </ErrorAlert>
+          )}
           <SessionEnvironmentFormContent control={control} errors={errors} />
         </ModalBody>
         <ModalFooter>
