+++ client/src/features/favicon/Favicon.tsx
@@ -18,7 +18,7 @@
 
 import { useEffect } from "react";
 import useAppSelector from "../../utils/customHooks/useAppSelector.hook";
-import { FAVICON_BY_SESSION_STATUS } from "../sessionsV2/session.utils";
+import { FAVICON_BY_SESSION_STATUS } from "../sessionsV2/session.constants";
 
 export function Favicon() {
   const favicon = useAppSelector(({ display }) => display.favicon);
